refactor(next-drupal): add getAuthorizationHeader method to DrupalClient
