+++ packages/next-drupal/tests/DrupalClient/basic-methods.test.ts
@@ -1,15 +1,10 @@
 import { afterEach, describe, expect, jest, test } from "@jest/globals"
-import { NextApiRequest, NextApiResponse } from "next"
-import { DRAFT_DATA_COOKIE_NAME, DrupalClient, JsonApiErrors } from "../../src"
+import { DrupalClient, JsonApiErrors } from "../../src"
 import { BASE_URL, mockLogger, spyOnFetch, spyOnFetchOnce } from "../utils"
 import type { DrupalNode, JsonApiError, Serializer } from "../../src"
 
 afterEach(() => {
   jest.restoreAllMocks()
-})
-
-describe("buildMenuTree()", () => {
-  test.todo("add tests")
 })
 
 describe("buildUrl()", () => {
@@ -240,199 +235,6 @@ describe("getErrorsFromResponse()", () => {
     })
 
     expect(await client.getErrorsFromResponse(response)).toBe("Forbidden")
-  })
-})
-
-describe("preview()", () => {
-  // Get values from our mocked request.
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  const { slug, resourceVersion, plugin, secret, ...draftData } =
-    new NextApiRequest().query
-  const dataCookie = `${DRAFT_DATA_COOKIE_NAME}=${encodeURIComponent(
-    JSON.stringify({ slug, resourceVersion, ...draftData })
-  )}; Path=/; HttpOnly; SameSite=None; Secure`
-  const validationPayload = {
-    slug,
-    maxAge: 30,
-  }
-
-  test("turns on preview mode and clears preview data", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const client = new DrupalClient(BASE_URL)
-    spyOnFetch({ responseBody: validationPayload })
-
-    await client.preview(request, response)
-
-    expect(response.clearPreviewData).toBeCalledTimes(1)
-    expect(response.setPreviewData).toBeCalledWith({
-      resourceVersion,
-      plugin,
-      ...validationPayload,
-    })
-  })
-
-  test("does not enable preview mode if validation fails", async () => {
-    const logger = mockLogger()
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const client = new DrupalClient(BASE_URL, { debug: true, logger })
-    const status = 403
-    const message = "mock fail"
-    spyOnFetch({
-      responseBody: { message },
-      status,
-      headers: {
-        "Content-Type": "application/json",
-      },
-    })
-
-    await client.preview(request, response)
-
-    expect(logger.debug).toBeCalledWith(
-      `Draft url validation error: ${message}`
-    )
-    expect(response.setPreviewData).toBeCalledTimes(0)
-    expect(response.statusCode).toBe(status)
-    expect(response.json).toBeCalledWith({ message })
-  })
-
-  test("does not turn on draft mode by default", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const client = new DrupalClient(BASE_URL)
-    spyOnFetch({ responseBody: validationPayload })
-
-    await client.preview(request, response)
-
-    expect(response.setDraftMode).toBeCalledTimes(0)
-
-    // Also check for no draft data cookie.
-    const cookies = response.getHeader("Set-Cookie")
-    expect(cookies[cookies.length - 1]).not.toBe(dataCookie)
-  })
-
-  test("optionally turns on draft mode", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const logger = mockLogger()
-    const client = new DrupalClient(BASE_URL, {
-      debug: true,
-      logger,
-    })
-    spyOnFetch({ responseBody: validationPayload })
-
-    const options = { enable: true }
-    await client.preview(request, response, options)
-
-    expect(response.setDraftMode).toBeCalledWith(options)
-
-    // Also check for draft data cookie.
-    const cookies = response.getHeader("Set-Cookie")
-    expect(cookies[cookies.length - 1]).toBe(dataCookie)
-
-    expect(logger.debug).toHaveBeenLastCalledWith("Draft mode enabled.")
-  })
-
-  test("updates preview mode cookie’s sameSite flag", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const client = new DrupalClient(BASE_URL)
-    spyOnFetch({ responseBody: validationPayload })
-
-    // Our mock response.setPreviewData() does not set a cookie, so we set one.
-    const previewCookie =
-      "__next_preview_data=secret-data; Path=/; HttpOnly; SameSite=Lax"
-    response.setHeader("Set-Cookie", [
-      previewCookie,
-      ...response.getHeader("Set-Cookie"),
-    ])
-
-    const cookies = response.getHeader("Set-Cookie")
-    cookies[0] = cookies[0].replace("SameSite=Lax", "SameSite=None; Secure")
-
-    await client.preview(request, response)
-
-    expect(response.getHeader).toHaveBeenLastCalledWith("Set-Cookie")
-    expect(response.setHeader).toHaveBeenLastCalledWith("Set-Cookie", cookies)
-    expect(response.getHeader("Set-Cookie")).toStrictEqual(cookies)
-  })
-
-  test("redirects to the slug path", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const logger = mockLogger()
-    const client = new DrupalClient(BASE_URL, { debug: true, logger })
-    spyOnFetch({ responseBody: validationPayload })
-
-    await client.preview(request, response)
-
-    expect(response.setPreviewData).toBeCalledWith({
-      resourceVersion,
-      plugin,
-      ...validationPayload,
-    })
-    expect(response.writeHead).toBeCalledWith(307, { Location: slug })
-    expect(logger.debug).toHaveBeenLastCalledWith("Preview mode enabled.")
-  })
-
-  test("returns a 422 response on error", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const logger = mockLogger()
-    const client = new DrupalClient(BASE_URL, { debug: true, logger })
-    const message = "mock internal error"
-    response.clearPreviewData = jest.fn(() => {
-      throw new Error(message)
-    })
-
-    await client.preview(request, response)
-
-    expect(logger.debug).toHaveBeenLastCalledWith(`Preview failed: ${message}`)
-    expect(response.status).toBeCalledWith(422)
-    expect(response.end).toHaveBeenCalled()
-  })
-})
-
-describe("previewDisable()", () => {
-  test("clears preview data", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const client = new DrupalClient(BASE_URL)
-
-    await client.previewDisable(request, response)
-    expect(response.clearPreviewData).toBeCalledTimes(1)
-  })
-
-  test("disables draft mode", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const client = new DrupalClient(BASE_URL)
-
-    await client.previewDisable(request, response)
-    expect(response.setDraftMode).toBeCalledWith({ enable: false })
-  })
-
-  test("deletes the draft cookie", async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const client = new DrupalClient(BASE_URL)
-
-    await client.previewDisable(request, response)
-    const cookies = response.getHeader("Set-Cookie")
-    expect(cookies[cookies.length - 1]).toBe(
-      `${DRAFT_DATA_COOKIE_NAME}=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT; HttpOnly; SameSite=None; Secure`
-    )
-  })
-
-  test('redirects to "/"', async () => {
-    const request = new NextApiRequest()
-    const response = new NextApiResponse()
-    const client = new DrupalClient(BASE_URL)
-
-    await client.previewDisable(request, response)
-    expect(response.writeHead).toBeCalledWith(307, { Location: "/" })
-    expect(response.end).toBeCalled()
   })
 })
