+++ packages/next-drupal/tests/DrupalClient/resource-methods.test.ts
@@ -1,343 +1,14 @@
 import { afterEach, describe, expect, jest, test } from "@jest/globals"
 import { DrupalClient } from "../../src"
-import {
-  BASE_URL,
-  mockLogger,
-  mocks,
-  spyOnFetch,
-  spyOnFetchOnce,
-} from "../utils"
-import type {
-  AccessToken,
-  DrupalClientAuth,
-  DrupalNode,
-  DrupalSearchApiJsonApiResponse,
-} from "../../src"
+import { BASE_URL, mocks, spyOnFetch } from "../utils"
+import type { DrupalNode, DrupalSearchApiJsonApiResponse } from "../../src"
 
 afterEach(() => {
   jest.restoreAllMocks()
 })
 
-describe("fetch()", () => {
-  test("allows fetching custom url", async () => {
-    const client = new DrupalClient(BASE_URL)
-    const url = client.buildUrl(
-      "/jsonapi/node/article/52837ad0-f218-46bd-a106-5710336b7053"
-    )
-
-    const response = await client.fetch(url.toString())
-    expect(response.headers.get("content-type")).toEqual(
-      "application/vnd.api+json"
-    )
-    const json = await response.json()
-    expect(json).toMatchSnapshot()
-  })
-
-  test("allows for custom fetcher", async () => {
-    const customFetch = jest.fn()
-
-    const client = new DrupalClient(BASE_URL, {
-      fetcher: customFetch,
-    })
-    const url = client.buildUrl("/jsonapi").toString()
-
-    await client.fetch(url)
-    expect(customFetch).toHaveBeenCalledWith(
-      url,
-      expect.objectContaining({
-        headers: {
-          "Content-Type": "application/vnd.api+json",
-          Accept: "application/vnd.api+json",
-        },
-      })
-    )
-
-    await client.fetch(url, {
-      headers: {
-        foo: "bar",
-      },
-    })
-    expect(customFetch).toHaveBeenLastCalledWith(
-      url,
-      expect.objectContaining({
-        headers: {
-          Accept: "application/vnd.api+json",
-          "Content-Type": "application/vnd.api+json",
-          foo: "bar",
-        },
-      })
-    )
-  })
-
-  describe("authentication", () => {
-    const clientIdSecret = mocks.auth.clientIdSecret
-
-    test("throws an error if withAuth is called when auth is not configured", async () => {
-      const client = new DrupalClient(BASE_URL)
-
-      const url = client.buildUrl("/jsonapi")
-
-      await expect(
-        client.fetch(url.toString(), {
-          withAuth: true,
-        })
-      ).rejects.toThrow("auth is not configured.")
-    })
-
-    test("accepts username and password", async () => {
-      const customFetch = jest.fn()
-
-      const client = new DrupalClient(BASE_URL, {
-        auth: {
-          username: "admin",
-          password: "password",
-        },
-        fetcher: customFetch,
-      })
-      const url = client.buildUrl("/jsonapi").toString()
-
-      await client.fetch(url, { withAuth: true })
-      expect(customFetch).toHaveBeenCalledWith(
-        url,
-        expect.objectContaining({
-          headers: {
-            "Content-Type": "application/vnd.api+json",
-            Accept: "application/vnd.api+json",
-            Authorization: "Basic YWRtaW46cGFzc3dvcmQ=",
-          },
-        })
-      )
-    })
-
-    test("accepts callback", async () => {
-      const customAuth = jest
-        .fn()
-        .mockReturnValue(
-          "Basic YXJzaGFkQG5leHQtZHJ1cGFsLm9yZzphYmMxMjM="
-        ) as DrupalClientAuth
-      const customFetch = jest.fn()
-
-      const client = new DrupalClient(BASE_URL, {
-        auth: customAuth,
-        fetcher: customFetch,
-      })
-      const url = client.buildUrl("/jsonapi").toString()
-
-      await client.fetch(url, { withAuth: true })
-      expect(customFetch).toHaveBeenCalledWith(
-        url,
-        expect.objectContaining({
-          headers: {
-            "Content-Type": "application/vnd.api+json",
-            Accept: "application/vnd.api+json",
-            Authorization: "Basic YXJzaGFkQG5leHQtZHJ1cGFsLm9yZzphYmMxMjM=",
-          },
-        })
-      )
-    })
-
-    test("accepts clientId and clientSecret", async () => {
-      const client = new DrupalClient(BASE_URL, {
-        auth: clientIdSecret,
-      })
-      const fetchSpy = spyOnFetch()
-
-      const basic = Buffer.from(
-        `${clientIdSecret.clientId}:${clientIdSecret.clientSecret}`
-      ).toString("base64")
-
-      await client.fetch("http://example.com", { withAuth: true })
-      expect(fetchSpy).toHaveBeenNthCalledWith(
-        1,
-        `${BASE_URL}/oauth/token`,
-        expect.objectContaining({
-          headers: {
-            Accept: "application/json",
-            Authorization: `Basic ${basic}`,
-            "Content-Type": "application/x-www-form-urlencoded",
-          },
-        })
-      )
-    })
-
-    test("accepts custom auth url", async () => {
-      const client = new DrupalClient(BASE_URL, {
-        auth: {
-          ...clientIdSecret,
-          url: "/custom/oauth",
-        },
-      })
-      const fetchSpy = spyOnFetch()
-
-      await client.fetch("http://example.com", { withAuth: true })
-      expect(fetchSpy).toHaveBeenNthCalledWith(
-        1,
-        `${BASE_URL}/custom/oauth`,
-        expect.anything()
-      )
-    })
-  })
-
-  describe("headers", () => {
-    // TODO: Are these duplicates of getters-setters/headers tests?
-    test("allows setting custom headers", async () => {
-      const customFetch = jest.fn()
-      const client = new DrupalClient(BASE_URL, {
-        fetcher: customFetch,
-      })
-      client.headers = {
-        foo: "bar",
-      }
-
-      const url = "http://example.com"
-
-      await client.fetch(url)
-      expect(customFetch).toHaveBeenCalledWith(
-        url,
-        expect.objectContaining({
-          headers: { foo: "bar" },
-        })
-      )
-    })
-
-    test("allows setting custom headers with custom auth", async () => {
-      const customFetch = jest.fn()
-      const client = new DrupalClient(BASE_URL, {
-        fetcher: customFetch,
-        headers: {
-          foo: "bar",
-        },
-        auth: jest
-          .fn()
-          .mockReturnValue(
-            "Basic YXJzaGFkQG5leHQtZHJ1cGFsLm9yZzphYmMxMjM="
-          ) as DrupalClientAuth,
-      })
-
-      const url = "http://example.com"
-
-      await client.fetch(url, { withAuth: true })
-
-      expect(customFetch).toHaveBeenCalledWith(
-        url,
-        expect.objectContaining({
-          headers: {
-            foo: "bar",
-            Authorization: "Basic YXJzaGFkQG5leHQtZHJ1cGFsLm9yZzphYmMxMjM=",
-          },
-        })
-      )
-    })
-  })
-})
-
-describe("getAccessToken()", () => {
-  const accessToken = mocks.auth.accessToken
-  const clientIdSecret = mocks.auth.clientIdSecret
-
-  test("uses the long-lived access token from constructor", async () => {
-    const longLivedAccessToken: AccessToken = {
-      ...accessToken,
-      expires_in: 360000,
-    }
-    const client = new DrupalClient(BASE_URL, {
-      accessToken: longLivedAccessToken,
-    })
-    const fetchSpy = spyOnFetch({
-      responseBody: {
-        ...accessToken,
-        access_token: "not-used",
-      },
-    })
-
-    const token = await client.getAccessToken({
-      clientId: "",
-      clientSecret: "",
-      scope: undefined,
-    })
-    expect(fetchSpy).toHaveBeenCalledTimes(0)
-    expect(token).toBe(longLivedAccessToken)
-  })
-
-  test("throws if auth is not configured", async () => {
-    const fetchSpy = spyOnFetch({
-      responseBody: accessToken,
-    })
-
-    const client = new DrupalClient(BASE_URL)
-
-    await expect(
-      // @ts-ignore
-      client.getAccessToken({ clientId: clientIdSecret.clientId })
-    ).rejects.toThrow(
-      "auth is not configured. See https://next-drupal.org/docs/client/auth"
-    )
-    expect(fetchSpy).toHaveBeenCalledTimes(0)
-  })
-
-  test("throws if auth is not ClientIdSecret", async () => {
-    const fetchSpy = spyOnFetch({
-      responseBody: accessToken,
-    })
-
-    const client = new DrupalClient(BASE_URL, {
-      auth: mocks.auth.basicAuth,
-      withAuth: true,
-    })
-
-    await expect(
-      // @ts-ignore
-      client.getAccessToken()
-    ).rejects.toThrow(
-      "'clientId' and 'clientSecret' required. See https://next-drupal.org/docs/client/auth"
-    )
-    expect(fetchSpy).toHaveBeenCalledTimes(0)
-  })
-
-  test("fetches an access token", async () => {
-    spyOnFetch({
-      responseBody: accessToken,
-    })
-
-    const logger = mockLogger()
-    const client = new DrupalClient(BASE_URL, {
-      auth: clientIdSecret,
-      debug: true,
-      logger,
-    })
-
-    const token = await client.getAccessToken()
-    expect(token).toEqual(accessToken)
-    expect(logger.debug).toHaveBeenCalledWith("Fetching new access token.")
-  })
-
-  test("re-uses access token", async () => {
-    spyOnFetchOnce({
-      responseBody: accessToken,
-    })
-    const fetchSpy = spyOnFetchOnce({
-      responseBody: {
-        ...accessToken,
-        access_token: "differentAccessToken",
-        expires_in: 1800,
-      },
-    })
-
-    const logger = mockLogger()
-    const client = new DrupalClient(BASE_URL, {
-      auth: clientIdSecret,
-      debug: true,
-      logger,
-    })
-
-    const token1 = await client.getAccessToken()
-    const token2 = await client.getAccessToken()
-    expect(token1).toEqual(token2)
-    expect(logger.debug).toHaveBeenLastCalledWith(
-      "Using existing access token."
-    )
-    expect(fetchSpy).toHaveBeenCalledTimes(1)
-  })
+describe("buildMenuTree()", () => {
+  test.todo("add tests")
 })
 
 describe("getEntryForResourceType()", () => {
